fixup! Configure the HTTP/2 connection window to 1MB

+++ linkerd/app/src/env.rs
@@ -181,8 +181,8 @@ const DEFAULT_OUTBOUND_CONNECT_BACKOFF: ExponentialBackoff = ExponentialBackoff 
 const DEFAULT_DNS_CANONICALIZE_TIMEOUT: Duration = Duration::from_millis(100);
 const DEFAULT_RESOLV_CONF: &str = "/etc/resolv.conf";
 
-const DEFAULT_INITIAL_STREAM_WINDOW_SIZE: usize = 65_535; // Protocol default
-const DEFAULT_INITIAL_CONNECTION_WINDOW_SIZE: usize = 1048576; // 1MB ~ 16 streams at capacity
+const DEFAULT_INITIAL_STREAM_WINDOW_SIZE: u32 = 65_535; // Protocol default
+const DEFAULT_INITIAL_CONNECTION_WINDOW_SIZE: u32 = 1048576; // 1MB ~ 16 streams at capacity
 
 /// It's assumed that a typical proxy can serve inbound traffic for up to 100 pod-local
 /// HTTP services and may communicate with up to 10K external HTTP domains.
